Add files via upload
got rid of brackets, quotes, commas, etc on variables leading up to confimation page, but the values are still in one big array and are displaying quotes. I am beyond frustrated trying to figure this out. It seems so simple and straight forward, but for some reason it has left me bumfuzzled.